+++ app/src/main/java/com/example/playlistmaker/domain/repository/FavoriteTracksRepository.kt
@@ -7,5 +7,5 @@ interface FavoriteTracksRepository {
     suspend fun addTrackToFavorites(track: Track)
     suspend fun removeTrackFromFavorites(track: Track)
     fun getFavoriteTracks(): Flow<List<Track>>
-    fun getFavoriteTrackIds(): Flow<List<Int>>
+    fun getFavoriteTrackIds(): Flow<List<Long>>
 }
